git add AssetManagement/client-app

+++ laravel/app/Http/Controllers/CategoryController.php
@@ -7,57 +7,48 @@ use App\Models\Category;
 
 class CategoryController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
-    public function index()
-    {
-        //
-        return Category::with('children')->get();
-
+    public function index(){
+        $categories = Category::all();
+        return response()->json($categories);
     }
-
-    /**
-     * Store a newly created resource in storage.
-     */
-    public function store(Request $request)
-    {
-        //
-        $request->validate(['name'=>'required|string']);
-        $category = Category::create([
-            'name'=>$request->name,
-            'parent_id'=>$request->parent_id
+    
+    public function create(Request $request){
+        $validated = $request->validate([
+            'name' => 'required|string|max:255',
+            'description' => 'nullable|string'
         ]);
-        return response()->json($category,201);
+
+        $category = Category::create($validated);
+
+        return response()->json($category, 201);    
     }
 
-    /**
-     * Display the specified resource.
-     */
-    public function show(string $id)
-    {
-        //
-    }
+    public function update(Request $request, $id){
 
-    /**
-     * Update the specified resource in storage.
-     */
-    public function update(Request $request, string $id)
-    {
-        //
-        $category = Category::findOrFail($id);
-        $category->update($request->only('name','parent_id'));
+        $category = Category::findorFail($id);
+        $validated = $request->validate([
+            'name'=> 'sometimes|required|string|max:255',
+            'description'=> 'nullable|string'
+        ]);
+
+        $category->update($validated);
         return response()->json($category);
-    }
+    } 
 
-    /**
-     * Remove the specified resource from storage.
-     */
-    public function destroy(string $id)
-    {
-        //
-        $category = Category::findOrFail($id);
+    public function destroy($id){
+
+        $category = Category::findorFail($id);
+
+         if ($category->assets()->exists() || $category->parameters()->exists()) {
+            return response()->json([
+                'message' => 'Cannot delete category with existing assets or parameters.'
+            ], 409);
+        }
+
         $category->delete();
-        return response()->json(['message'=>'Deleted']);
+
+        return response()->json([
+            'message' => 'Category deleted successfully'
+        ]);
     }
 }
